% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTN-package.R
\docType{package}
\name{BTN-package}
\alias{BTN}
\alias{BTN-package}
\title{BTN: Helper functions to support use of the Bergen Telemetry Network}
\description{
Provides access to shapefiles for Aurland and Vosso. Imports data from Thelma tbdb databases and helps match detections to receiver and tag metadata stored in google sheets.
}
\author{
\strong{Maintainer}: Robert Lennox \email{lennox@dal.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
