% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{gantt}
\alias{gantt}
\title{functions to get receiver and tagging metadata into session}
\arguments{
\item{url}{is the google sheet loaded by gsheet::gsheet2tbl function}

\item{y}{is the starting year of the project}

\item{m}{is the starting month of the project}

\item{slice}{is the number of rows to use in the gantt chart}
}
\description{
functions to get receiver and tagging metadata into session
}
