% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{smoltify}
\alias{smoltify}
\title{functions to get receiver and tagging metadata into session}
\arguments{
\item{meta}{is an object brought in from the tagging metadata using the gsheet::gsheet2tbl function}

\item{receivers}{is an object brought in from the receivers sheet using the gsheet::gsheet2tbl function}
}
\description{
functions to get receiver and tagging metadata into session
}
