% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaps_all.R
\name{yaps_all}
\alias{yaps_all}
\title{function to read in and manipulate Thelma DB}
\usage{
yaps_all(
  fish_detections,
  sync_model,
  hydros,
  transmitter_ID,
  rbi_min = 60,
  rbi_max = 120,
  nruns = 5
)
}
\arguments{
\item{sync_model}{is the chosen synchronization model from YAPS}

\item{hydros}{is the formatted hydrophone data.table}

\item{transmitter_ID}{is the transmitter ID desired to be positioned}

\item{rbi_min}{is the minimum time of the tag transmissions, default is 60}

\item{rbi_max}{is the maximim time of the tag transmissions, default is 120}

\item{nruns}{is the number of YAPS fits desired, default is 5}

\item{fish_detecions}{is the object containing the YAPS formatted fish detections}
}
\description{
function to read in and manipulate Thelma DB
}
\examples{
yaps_all(fish_detections=dets, sync_model=sync, transmitter_ID=4000)

}
