% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb.R
\name{hatch}
\alias{hatch}
\title{function to import data from Thelma tbdb}
\usage{
hatch(tbdb_file)
}
\arguments{
\item{tbdb_file}{is the name of the thelma database e.g. "PACE.tbdb"}
}
\description{
function to import data from Thelma tbdb
}
