% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tb.R
\name{hatch}
\alias{hatch}
\title{function to read in and manipulate Thelma DB}
\usage{
hatch(tbdb_file)
}
\arguments{
\item{tbdb_file}{is the name of the thelma database}
}
\description{
function to read in and manipulate Thelma DB
}
\examples{
hatch(tbdb_file) where tbdb_file is the file path to a thelma database
e.g. "PACE.tbdb"

}
